+++ tsconfig.json
@@ -5,8 +5,8 @@
   "exclude": [
     "./node_modules/",
     "genenv.js",
-    "./src/commands/orders/worker/ebrew.ts",
-    "./src/commands/orders/worker/eclaim.ts"
+    "./purgatory/ebrew.ts",
+    "./purgatory/eclaim.ts"
   ],
   "compilerOptions": {
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
